requirements.txt
beautifulsoup4==4.12.2
pwntools==4.10.0
requests==2.31.0